requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=1.4
asyncpg
aiosqlite
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt>=4.0.1
python-multipart
python-dotenv
alembic
pytest
httpx
pytest-asyncio
psycopg2-binary
loguru
pydantic[email]
requests